backend/app/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2-binary>=2.8.6,<2.9
postgis
googlemaps>=4.10.0,<4.11


